backend/pyproject.toml
[tool.poetry]
name = "piecewise"
version = "0.1.0"
description = "A tool for digesting and aggregating user-volunteered Internet performance test results data from Measurement Lab, and visualizing aggregate data on the web."
authors = ["Josh King <josh@throneless.tech>, Rae Gaines <rae@throneless.tech>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.45.0"
sqlalchemy = "^1.3.11"
ujson = "^1.35"
graphene = "^2.1.8"
aiofiles = "^0.4.0"
uvicorn = "^0.10.8"
jinja2 = "^2.10.3"
python-multipart = "^0.0.5"
alembic = "^1.3.2"
email-validator = "^1.0.5"
passlib = "^1.7.2"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]
requests = "^2.22.0"
pytest = "^5.3.2"
pylint = "^2.4.4"
portray = "^1.3.1"
pdocs = "^1.0.1"
tox = "^3.14.2"
pytest-cov = "^2.8.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
